Moved options into constants
