+++ MyMoney.Web/ClientApp/src/app/pages/incomes/edit-income/edit-recurring-income/edit-recurring-income.component.ts
@@ -3,7 +3,8 @@ import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { Router, ActivatedRoute } from '@angular/router';
 import { IncomeService } from '../../../../shared/services';
 import { IRecurringIncomeDto, Frequency, IIncomeDto } from 'src/app/shared/api';
-import { frequencyOptions, toInputDateString } from 'src/app/shared/functions';
+import { toInputDateString } from 'src/app/shared/functions';
+import { frequencyOptions } from 'src/app/shared/constants';
 
 @Component({
    templateUrl: './edit-recurring-income.component.html',
@@ -18,16 +19,14 @@ export class EditRecurringIncomeComponent implements OnInit {
    public submitted = false;
    public isValid = true;
    public children: { id: number; date: string }[] = [];
-   public recurrenceOptions: { key: Frequency; value: string }[];
+   public recurrenceOptions: { key: Frequency; value: string }[] = frequencyOptions;
 
    constructor(
       private readonly formBuilder: FormBuilder,
       private readonly incomeService: IncomeService,
       private readonly router: Router,
       private readonly activatedRoute: ActivatedRoute,
-   ) {
-      this.recurrenceOptions = frequencyOptions;
-   }
+   ) { }
 
    public ngOnInit(): void {
