+++ MyMoney.Web/ClientApp/src/app/pages/transactions/add-transaction/add-recurring-transaction/add-recurring-transaction.component.ts
@@ -3,7 +3,7 @@ import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { Router } from '@angular/router';
 import { TransactionService } from 'src/app/shared/services';
 import { IRecurringTransactionDto, Frequency } from 'src/app/shared/api';
-import { frequencyOptions } from 'src/app/shared/functions';
+import { frequencyOptions } from 'src/app/shared/constants';
 
 @Component({
    selector: 'mymoney-add-recurring-transaction',
@@ -14,15 +14,13 @@ export class AddRecurringTransactionComponent implements OnInit {
    public addTransactionForm: FormGroup;
    public loading = false;
    public submitted = false;
-   public recurrenceOptions: { key: Frequency; value: string }[];
+   public recurrenceOptions: { key: Frequency; value: string }[] = frequencyOptions;
 
    constructor(
       private readonly formBuilder: FormBuilder,
       private readonly router: Router,
       private readonly transactionService: TransactionService,
-   ) {
-      this.recurrenceOptions = frequencyOptions;
-   }
+   ) { }
 
    public ngOnInit(): void {
