+++ MyMoney.Web/ClientApp/src/app/shared/functions.ts
@@ -40,12 +40,3 @@ export function toFrequencyString(frequency: Frequency): string {
       case Frequency.year: return 'Annually';
    }
 }
-
-export const frequencyOptions: { key: Frequency; value: string }[] = [
-   { key: Frequency.day, value: toFrequencyString(Frequency.day) },
-   { key: Frequency.week, value: toFrequencyString(Frequency.week) },
-   { key: Frequency.fortnight, value: toFrequencyString(Frequency.fortnight) },
-   { key: Frequency.fourWeek, value: toFrequencyString(Frequency.fourWeek) },
-   { key: Frequency.month, value: toFrequencyString(Frequency.month) },
-   { key: Frequency.year, value: toFrequencyString(Frequency.year) },
-];
